v0.0.2 VLA C3: multi-dimensional VLAs — int a[n][m] RUNS with a runtime inner stride on all 3 legs
The variably-modified-type step of the VLA arc: every block-scope multi-dim VLA
shape now runs on all 3 legs (LEAF scope) with a RUNTIME inner stride —
`int a[n][m]`, `int a[n][m][k]`, `int a[n][5]`, and `int a[5][n]` (fixed outer).
`c99_vla_multidim` runs and returns 42 on x86_64-elf, arm64-elf (qemu), and pe64.

§B DECIDED (user 2026-07-13): storage = Option 1 (store cumulative row-strides ONCE
at the declaration, Load at each index); `int a[5][n]` INCLUDED (no-compromise) via a
transitive `typeContainsVla`.

Single-tier — no LIR/asm/target/format change (the dynamic alloca, frame pointer, and
byte-Gep all exist from C1b; a runtime stride is just more MIR Mul/Load). Mechanism:
- Nested `vlaArray(vlaArray(int))` type — the fixed multi-dim `Index(Index(...))` +
  `indexResultType`-peel + byte-Gep machinery composes for free (a VLA is kind Array
  everywhere). The four `S_VlaMultiDimUnsupported` rejects are lifted.
- `interner.typeContainsVla` (recursive ops[0]/isVlaArray walk; does NOT over-fire on a
  fixed `int[5][5]`) threads 8 `isVlaArray` dispatch gates so a top-fixed `a[5][n]`
  routes to the VLA path (the 4 rejects, validateVlaDeclarator, the captureVlaSize gate,
  the sizeof recording, and the MIR verifier's alloca-operand rule).
- `captureVlaSize` collects EVERY declarator suffix (a vector, outer->inner);
  `vlaAllocaForLocal` evaluates all dims exactly once at the decl (C 6.7.6.2p2,
  side-effect-safe for `int a[(k+=3)][(k+=5)]`), folds cumulative row-strides bottom-up,
  and freezes one slot per runtime-sized level keyed by (baseSymbol, elemTy TypeId).

THE CRITICAL-1 CONVENTION (a self-contradictory stride-level map — caught by the
design-audit before any code — would have miscompiled `a[1][0]` into `a[0][1]`): scale a
subscript / size a row by the runtime `sizeof` of the SUBSCRIPT-RESULT type. Keying the
slots by (baseSymbol, elemTy TypeId) makes the store key at the decl and the load key at
each index/sizeof the SAME interned `operands()` peel — they match by construction, no
depth arithmetic. This unifies the outer index, `sizeof a[0]` (a runtime row), and
`sizeof a` (the whole, = the C2 total slot) into one stored family. A fixed row
(`int a[n][5]`'s `int[5]`) is not slotted — it uses the compile-time `elementStride`.

Independent design-audit GO-WITH-FIXES folded 3 CRITICALs before coding: the stride-level
map, the MIR verifier rejecting `a[5][n]`'s runtime-operand alloca, and 4 more
`isVlaArray` gates `a[5][n]` would have leaked past. Independent code-audit: SHIP (stride
arithmetic hand-verified per shape; store/load keys match; evaluate-once holds;
typeContainsVla complete + non-over-firing) — folded one MINOR: the misleading "internal
side-table desync" fail-loud message for the C4-deferred shapes (a VLA whose element is
an array typedef; a pointer-to-VLA) is reworded to an honest unsupported-shape diagnostic.

Orchestrator-verified independently: the off-diagonal `a[1][0]` (asymmetric dims) reads
back 77, not the transposed 88, on all 3 legs — the definitive stride-order check;
`sizeof a`/`sizeof a[0]`/`sizeof a[0][0]` correct; per-dim non-integer -> S_VlaSizeNotInteger
(fires on the offending dim); file-scope multi-dim VLA -> S_NonConstantArrayLength;
wide `_BitInt(N>64)` bound -> H0009; struct-member VLA -> S_FlexibleArraySoleMember;
`_Static_assert(sizeof a == K)` on a multi-dim VLA -> S0029 (the C2 invariant preserved);
a fixed `int b[5][5]` does not route to the VLA path.

`S_VlaMultiDimUnsupported` (0xE052) is now unemitted but kept, reserved for the C4 `[*]`
shape. ptr-to-VLA / VLA typedef / `[*]` / `[static n]` remain C4 (fail loud today).

Tests: runnable `c99_vla_multidim` (4-target + release arm; asymmetric off-diagonal
write, 3-D, `a[5][n]`, evaluate-once, in-program sizeof pins; exit 42); MIR pins
`MultiDimVlaLowersWithRuntimeStrideSlots`, `FixedMultiDimArrayDoesNotRouteToVlaPath` (a
direct `typeContainsVla` interner pin), `RowSizeofOfMultiDimVlaLoadsStrideSlot`; corpus
goldens repurposed (per-dim non-integer, file-scope routing) + a struct-member reject.

Gates: Windows Debug ctest 633/633, WSL x86_64 633/633, c99_vla_multidim RUNS -> 42 on
x86_64-elf + arm64-elf (qemu, debug+release) + pe64; sqlite 3.54.0 / CRUD 42 on all 3
legs; anchor guard OK; agnosticism clean (no arch/format identity in the stride/index/
typeContainsVla path). D-CSUBSET-VLA advances C2 -> C3. Next: C4 (pointer-to-VLA
`int (*p)[n]` + VLA typedef + `[*]` / `[static n]`).

+++ tests/corpus/diagnostics/c-subset/vla_multidim.c
@@ -1,8 +1,9 @@
-// VLA C1a (D-CSUBSET-VLA, MINOR-2): a multi-dimensional VLA `int a[n][m]` needs a
-// runtime STRIDE through index/GEP — the C3 boundary. The inner `[m]` folds to a
-// vlaArray; the outer `[n]` VLA arm rejects a VLA/array element loud
-// (S_VlaMultiDimUnsupported), never a silent nested-VLA.
-int f(int n, int m) {
+// VLA C3 (D-CSUBSET-VLA, IMPORTANT-7): a MULTI-DIMENSIONAL VLA now RUNS (`int a[n][m]`
+// lowers to a runtime row stride) — but EVERY dimension's bound must still have integer
+// type (C 6.7.6.2p1). The per-dim validator enumerates ALL suffixes, so a NON-integer
+// INNER dim (`[m]`, where `m` is float) is rejected ON THAT DIM (S_VlaSizeNotInteger),
+// never silently accepted. Red-on-regression for the multi-dim per-dim check.
+int f(int n, float m) {
     int a[n][m];
     return 0;
 }
